quantizer: rename package to obfuscate
