--- quantizer/cassandra_test.go
@@ -1,59 +1,0 @@
-package quantizer
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-
-	"github.com/DataDog/datadog-trace-agent/model"
-)
-
-func CassSpan(query string) *model.Span {
-	return &model.Span{
-		Resource: query,
-		Type:     "cassandra",
-		Meta: map[string]string{
-			"query": query,
-		},
-	}
-}
-
-func TestCassQuantizer(t *testing.T) {
-	assert := assert.New(t)
-
-	queryToExpected := []struct{ in, expected string }{
-		// List compacted and replaced
-		{
-			"select key, status, modified from org_check_run where org_id = %s and check in (%s,%s,%s,%s,%s,%s,%s,%s,%s,%s,%s)",
-			"select key, status, modified from org_check_run where org_id = ? and check in ( ? )",
-		},
-		// Some whitespace-y things
-		{
-			"select key, status, modified from org_check_run where org_id = %s and check in (%s, %s, %s)",
-			"select key, status, modified from org_check_run where org_id = ? and check in ( ? )",
-		},
-		{
-			"select key, status, modified from org_check_run where org_id = %s and check in (%s , %s , %s )",
-			"select key, status, modified from org_check_run where org_id = ? and check in ( ? )",
-		},
-		// %s replaced with ? as in sql quantize
-		{
-			"select key, status, modified from org_check_run where org_id = %s and check = %s",
-			"select key, status, modified from org_check_run where org_id = ? and check = ?",
-		},
-		{
-			"select key, status, modified from org_check_run where org_id = %s and check = %s",
-			"select key, status, modified from org_check_run where org_id = ? and check = ?",
-		},
-		{
-			"SELECT timestamp, processes FROM process_snapshot.minutely WHERE org_id = ? AND host = ? AND timestamp >= ? AND timestamp <= ?",
-			"SELECT timestamp, processes FROM process_snapshot.minutely WHERE org_id = ? AND host = ? AND timestamp >= ? AND timestamp <= ?",
-		},
-	}
-
-	for _, testCase := range queryToExpected {
-		s := CassSpan(testCase.in)
-		NewObfuscator(nil).Obfuscate(s)
-		assert.Equal(testCase.expected, s.Resource)
-	}
-}
